requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
coinapi.rest.v1==1.1
Django==2.2.7
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
websocket-client==0.56.0
zipp==0.6.0
